extract render callback into FormModalForm to run useEffect js_only
